Added the possibility to choose the type of connection security
This actually fixes several problems. Previously only choosing between secure and not was possible. But nodemailer still tried to establish a secure connection even when no security was chosen. In those cases nodemailer tried to use STARTTLS and failed when there was an invalid certificate. In this new version you can choose what type of security you want and nodemailer will respect that decision.